Reorganized the file structure. Improved mobile.
Now its easier to develop different aspects of the project, if you want to change something in relation to nodes you can go to ./methods/nodeMethods.js or linkMethods.js for links. This makes the code easier to read and update as each file has less lines to read. Added some mobile functionality that when you click the zoom out button, the selected html disappears. I made the click to focus method faster and not zoom in as much as before, I removed the toggle node names button as its not functional and I added an icon.

I still need to fix the show and hide arrow on the left of the search bar, improve search ease, and add a link search bar too.